requirements.txt
# === Django Core ===
Django==5.2.4
gunicorn==21.2.0
python-dotenv==1.2.1

# === AI & Embedding ===
sentence-transformers==5.0.0
transformers==4.53.2
torch==2.7.1
faiss-cpu==1.12.0
scikit-learn==1.7.0
scipy==1.15.3
numpy==1.26.4
pandas==2.3.1

# === Google Gemini API ===
google-generativeai==0.8.5
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-api-core==2.26.0

# === NLP & Web Parsing ===
ckip-transformers==0.3.4
beautifulsoup4==4.13.4
lxml==6.0.0
requests==2.32.4

# === Utilities ===
tqdm==4.67.1
regex==2024.11.6
typing-extensions==4.14.1
colorama==0.4.6
idna==3.10
urllib3==2.5.0
tzdata==2025.2
sqlparse==0.5.3